更新播放速度控制方式

+++ Assets/Scripts/UIScript/StudyMode/StudyModeManager.cs
@@ -44,7 +44,7 @@ public class StudyModeManager : MonoBehaviour
             m_RefRateController.nCurrentFrame += 1;
             m_RateController.nCurrentFrame += 1;
 
-            m_StudyController.Update(m_RefRateController.nCurrentFrame, m_RateController.nCurrentFrame);
+            m_StudyController.Update((int)m_RefRateController.nCurrentFrame, (int)m_RateController.nCurrentFrame);
             m_StudyModeChartController.UpdateRefLineChart(ChartType.CHART_SPEED, m_RefRateController.nCurrentFrame, TrailCurveDrawCtrl.Instance().lastSpeed(TrailType.EG_S1));
             m_StudyModeChartController.UpdateRefLineChart(ChartType.CHART_ACCELERATE, m_RefRateController.nCurrentFrame, TrailCurveDrawCtrl.Instance().lastAcceleration(TrailType.EG_S1));
             m_StudyModeChartController.UpdateRefLineChart(ChartType.CHART_CURVATURE, m_RefRateController.nCurrentFrame, TrailCurveDrawCtrl.Instance().lastCurvature(TrailType.EG_S1));
